+++ crates/task-impls/src/quorum_proposal/handlers.rs
@@ -26,7 +26,7 @@ use hotshot_types::{
     },
 };
 use tracing::instrument;
-use utils::result12345::*;
+use utils::anytrace::*;
 use vbs::version::StaticVersionType;
 
 use crate::{
