+++ crates/task-impls/src/quorum_proposal/mod.rs
@@ -33,7 +33,7 @@ use hotshot_types::{
 #[cfg(async_executor_impl = "tokio")]
 use tokio::task::JoinHandle;
 use tracing::instrument;
-use utils::result12345::*;
+use utils::anytrace::*;
 
 use self::handlers::{ProposalDependency, ProposalDependencyHandle};
 use crate::{
