+++ crates/task-impls/src/quorum_proposal_recv/handlers.rs
@@ -26,7 +26,7 @@ use hotshot_types::{
     vote::{Certificate, HasViewNumber},
 };
 use tracing::instrument;
-use utils::result12345::*;
+use utils::anytrace::*;
 
 use super::QuorumProposalRecvTaskState;
 use crate::{
