+++ crates/task-impls/src/quorum_vote/handlers.rs
@@ -19,7 +19,7 @@ use hotshot_types::{
     vote::HasViewNumber,
 };
 use tracing::instrument;
-use utils::result12345::*;
+use utils::anytrace::*;
 
 use super::QuorumVoteTaskState;
 use crate::{
