+++ crates/task-impls/src/quorum_vote/mod.rs
@@ -38,7 +38,7 @@ use jf_vid::VidScheme;
 #[cfg(async_executor_impl = "tokio")]
 use tokio::task::JoinHandle;
 use tracing::instrument;
-use utils::result12345::*;
+use utils::anytrace::*;
 
 use crate::{
     events::HotShotEvent,
