+++ crates/task-impls/src/rewind.rs
@@ -10,7 +10,7 @@ use async_broadcast::{Receiver, Sender};
 use async_trait::async_trait;
 use hotshot_task::task::TaskState;
 use hotshot_types::traits::node_implementation::NodeType;
-use utils::result12345::Result;
+use utils::anytrace::Result;
 
 use crate::events::HotShotEvent;
