+++ crates/task/src/task.rs
@@ -16,7 +16,7 @@ use futures::future::join_all;
 use futures::future::try_join_all;
 #[cfg(async_executor_impl = "tokio")]
 use tokio::task::{spawn, JoinHandle};
-use utils::result12345::Result;
+use utils::anytrace::Result;
 
 /// Trait for events that long-running tasks handle
 pub trait TaskEvent: PartialEq {
