+++ crates/types/src/data.rs
@@ -30,7 +30,7 @@ use thiserror::Error;
 #[cfg(async_executor_impl = "tokio")]
 use tokio::task::spawn_blocking;
 use tracing::error;
-use utils::result12345::*;
+use utils::anytrace::*;
 use vec1::Vec1;
 
 use crate::{
