+++ crates/types/src/message.rs
@@ -19,7 +19,7 @@ use async_lock::RwLock;
 use cdn_proto::util::mnemonic;
 use derivative::Derivative;
 use serde::{de::DeserializeOwned, Deserialize, Serialize};
-use utils::result12345::*;
+use utils::anytrace::*;
 use vbs::{
     version::{StaticVersionType, Version},
     BinarySerializer, Serializer,
