+++ crates/types/src/simple_certificate.rs
@@ -17,7 +17,7 @@ use async_lock::RwLock;
 use committable::{Commitment, Committable};
 use ethereum_types::U256;
 use serde::{Deserialize, Serialize};
-use utils::result12345::*;
+use utils::anytrace::*;
 
 use crate::{
     data::serialize_signature2,
