+++ crates/utils/src/lib.rs
@@ -1,4 +1,4 @@
 //! General (not HotShot-specific) utilities
 
 /// Error utilities, intended to function as a replacement to `anyhow`.
-pub mod result12345;
+pub mod anytrace;
